Refactor main function
